python/requirements.txt
annotated-types==0.7.0
asttokens==2.4.1
backcall==0.2.0
blis==0.7.11
catalogue==2.0.10
Cerberus==1.3.5
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
decorator==5.1.1
docopt==0.6.2
executing==2.0.1
filelock==3.9.0
flatbuffers==24.3.25
fonttools==4.45.1
fsspec==2023.4.0
gensim==4.3.2
h5py==3.11.0
idna==3.7
importlib-resources==6.1.1
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.0
nltk==3.8.1
num2words==0.5.13
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
onnx==1.16.1
onnxruntime-training==1.18.0
onnxscript==0.1.0.dev20240612
packaging==23.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==5.27.1
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
thinc==8.2.4
threadpoolctl==3.5.0
torch==2.3.1+cu118
torch-ort==1.18.0
torchaudio==2.3.1+cu118
torchvision==0.18.1+cu118
tqdm==4.66.4
traitlets==5.14.3
triton==2.3.1
typer==0.12.3
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wheel==0.43.0
wrapt==1.16.0
zipp==3.17.0
